Refactoring step 1: oshi.jna.platform packages
