+++ src/main/java/oshi/jna/platform/windows/Pdh.java
@@ -14,7 +14,7 @@
  * widdis[at]gmail[dot]com
  * https://github.com/dblock/oshi/graphs/contributors
  */
-package oshi.software.os.windows.nt;
+package oshi.jna.platform.windows;
 
 import java.util.Arrays;
 import java.util.List;
