+++ src/main/java/oshi/jna/platform/windows/package-info.java
@@ -1,0 +1,17 @@
+/**
+ * Oshi (https://github.com/dblock/oshi)
+ *
+ * Copyright (c) 2010 - 2015 The Oshi Project Team
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ * https://github.com/dblock/oshi/graphs/contributors
+ */
+/**
+ * Provides extensions of JNA libraries for Windows
+ */
+package oshi.jna.platform.windows;
