+++ src/main/java/oshi/software/os/linux/proc/CentralProcessor.java
@@ -31,8 +31,8 @@ import org.slf4j.LoggerFactory;
 import com.sun.jna.Native;
 
 import oshi.hardware.Processor;
-import oshi.software.os.linux.Libc;
-import oshi.software.os.linux.Libc.Sysinfo;
+import oshi.jna.platform.linux.Libc;
+import oshi.jna.platform.linux.Libc.Sysinfo;
 import oshi.util.ExecutingCommand;
 import oshi.util.FileUtil;
 import oshi.util.ParseUtil;
